functions/requirements.txt
firebase-functions>=0.4.0
firebase-admin>=6.4.0
requests>=2.31.0
rapidocr-onnxruntime>=1.3.0
Pillow>=10.0.0
flask>=3.0.0
python-dotenv>=1.0.0
